refactor TacticalSuperiorityPatch into a base class named SimulateHitPatch
